feat: detect connection errors and force a reconnect
Connections to SFTP servers are often idle for a long period
of time as jobs are ran on a schedule (e.g. hourly). This can cause
connections to be closed by a middleware, server, or other means.

Errors like `connection lost` are common in this scenario and while
callers of go-sftp could close connections after jobs finish and
reconnect at the start of jobs we can check for these errors and
reconnect automatically on the next call.